moved grid code into new worlds file
